third_party/WebKit/LayoutTests/webaudio/AudioBufferSource/audiobuffersource-loop-grain-no-duration-expected.txt

Test AudioBufferSourceNode looping without explicit duration

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS The output of actual and expected loops is identical to the array [0,0.125,0.25,0.375,0.5,0.625,0.75,0.875,0,0.125,0.25,0.375,0.5,0.625,0.75,0.875,...].
PASS The content of the left and right channel is identical to the array [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,...].
PASS successfullyParsed is true

TEST COMPLETE
